requirements.txt
customtkinter
darkdetect
numpy
packaging
pillow
